+++ weblate/trans/tests/test_create.py
@@ -53,7 +53,6 @@ class CreateTest(ViewTestCase):
             "name": "Create Project",
             "slug": "create-project",
             "web": "https://weblate.org/",
-            "source_language": get_english_lang(),
         }
         params.update(kwargs)
         response = self.client.post(reverse("create-project"), params)
@@ -119,6 +118,7 @@ class CreateTest(ViewTestCase):
             "new_base": "po/project.pot",
             "new_lang": "add",
             "language_regex": "^[^.]+$",
+            "source_language": get_english_lang(),
         }
         params.update(kwargs)
         response = self.client.post(reverse("create-component-vcs"), params)
@@ -178,6 +178,7 @@ class CreateTest(ViewTestCase):
             "project": self.project.pk,
             "vcs": "git",
             "repo": self.component.repo,
+            "source_language": get_english_lang(),
         }
         response = self.client.post(reverse("create-component-vcs"), params)
         self.assertContains(response, self.component.get_repo_link_url())
@@ -261,6 +262,7 @@ class CreateTest(ViewTestCase):
                     "name": "Create Component",
                     "slug": "create-component",
                     "project": self.project.pk,
+                    "source_language": get_english_lang(),
                 },
             )
         self.assertContains(response, "Failed to parse uploaded ZIP file.")
@@ -277,6 +279,7 @@ class CreateTest(ViewTestCase):
                     "name": "Create Component",
                     "slug": "create-component",
                     "project": self.project.pk,
+                    "source_language": get_english_lang(),
                 },
             )
         self.assertContains(response, "*.po")
@@ -290,6 +293,7 @@ class CreateTest(ViewTestCase):
                 "vcs": "local",
                 "repo": "local:",
                 "discovery": "0",
+                "source_language": get_english_lang(),
             },
         )
         self.assertContains(response, "Adding new translation")
@@ -307,6 +311,7 @@ class CreateTest(ViewTestCase):
                     "name": "Create Component",
                     "slug": "create-component",
                     "project": self.project.pk,
+                    "source_language": get_english_lang(),
                 },
             )
         self.assertContains(response, "*.html")
@@ -320,6 +325,7 @@ class CreateTest(ViewTestCase):
                 "vcs": "local",
                 "repo": "local:",
                 "discovery": "0",
+                "source_language": get_english_lang(),
             },
         )
         self.assertContains(response, "Adding new translation")
@@ -336,6 +342,7 @@ class CreateTest(ViewTestCase):
                     "slug": "create-component",
                     "project": self.project.pk,
                     "file_format": "po-mono",
+                    "source_language": get_english_lang(),
                 },
                 follow=True,
             )
@@ -364,6 +371,7 @@ class CreateTest(ViewTestCase):
                 "slug": "create-component",
                 "project": self.project.pk,
                 "file_format": "aresource",
+                "source_language": get_english_lang(),
             },
             follow=True,
         )
@@ -383,6 +391,7 @@ class CreateTest(ViewTestCase):
                 "slug": "create-component",
                 "project": self.project.pk,
                 "file_format": "po",
+                "source_language": get_english_lang(),
             },
             follow=True,
         )
@@ -402,6 +411,7 @@ class CreateTest(ViewTestCase):
                 "slug": "create-component",
                 "project": self.project.pk,
                 "file_format": "strings",
+                "source_language": get_english_lang(),
             },
             follow=True,
         )